Moving some files to correct directory
