+++ baylei-start/pom.xml
@@ -14,10 +14,6 @@
     <name>baylei-start</name>
     <description>Demo project for Spring Boot</description>
 
-<!--    <properties>-->
-<!--        <java.version>11</java.version>-->
-<!--    </properties>-->
-
     <dependencies>
         <dependency>
             <groupId>br.com.baylei</groupId>
@@ -38,13 +34,14 @@
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-mongodb</artifactId>
+            <artifactId>spring-boot-starter-actuator</artifactId>
         </dependency>
 
         <dependency>
-            <groupId>org.springframework.data</groupId>
-            <artifactId>spring-data-mongodb</artifactId>
+            <groupId>org.springframework.cloud</groupId>
+            <artifactId>spring-cloud-starter-config</artifactId>
         </dependency>
+
     </dependencies>
 
     <build>
